Implement inventory as an Observable Collection
When an item is added or removed from the collection
a CollectionChanged event if fired.
This implementation was made in order to have a way of updating
an UI element with items bound to the collection.
This collection might be used in any situation where
knowing if the collection has been changed matters.